// WADNR.Web/src/app/shared/generated/model/location-detailed-step.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectLocationItem } from './project-location-item';
import { ProjectLocationStagingItem } from './project-location-staging-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class LocationDetailedStep { 
    ProjectID?: number;
    Locations?: Array<ProjectLocationItem> | null;
    StagedLocations?: Array<ProjectLocationStagingItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface LocationDetailedStepForm { 
    ProjectID?: FormControl<number>;
    Locations?: FormControl<Array<ProjectLocationItem>>;
    StagedLocations?: FormControl<Array<ProjectLocationStagingItem>>;
}

export class LocationDetailedStepFormControls { 
    public static ProjectID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Locations = (value: FormControlState<Array<ProjectLocationItem>> | Array<ProjectLocationItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectLocationItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static StagedLocations = (value: FormControlState<Array<ProjectLocationStagingItem>> | Array<ProjectLocationStagingItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectLocationStagingItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}