// WADNR.Web/src/app/shared/generated/model/location-detailed-step-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LocationDetailedItemRequest } from './location-detailed-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class LocationDetailedStepRequest { 
    Locations?: Array<LocationDetailedItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface LocationDetailedStepRequestForm { 
    Locations?: FormControl<Array<LocationDetailedItemRequest>>;
}

export class LocationDetailedStepRequestFormControls { 
    public static Locations = (value: FormControlState<Array<LocationDetailedItemRequest>> | Array<LocationDetailedItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<LocationDetailedItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}