// WADNR.Web/src/app/shared/generated/model/location-detailed-item-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class LocationDetailedItemRequest { 
    ProjectLocationID?: number | null;
    ProjectLocationTypeID?: number;
    ProjectLocationNotes?: string | null;
    ProjectLocationName?: string | null;
    GeoJson?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface LocationDetailedItemRequestForm { 
    ProjectLocationID?: FormControl<number>;
    ProjectLocationTypeID?: FormControl<number>;
    ProjectLocationNotes?: FormControl<string>;
    ProjectLocationName?: FormControl<string>;
    GeoJson?: FormControl<string>;
}

export class LocationDetailedItemRequestFormControls { 
    public static ProjectLocationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectLocationTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectLocationNotes = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectLocationName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GeoJson = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}